requirements.txt
streamlit
openai
load_dotenv